HFP: Include retry logic when the device is disconnected once
When the device is disconnected once, the phone tries to
reconnect to the remote device once again after 2.5sec.
If the connection to the remote device is successful,
the entry of the remote is removed.

Change-Id: I6c040a95d9975e45cd844e25e474bc389be19180

Adding a new BroadcastReceiver for phone state change

When the SIM state is changed i.e loaded or unloaded,
the simStateChanged intent is received.
The  BroadcastReceiver mPhoneStateChangeReceiver
is registered during bluetooth ON and unregistered
during bluetooth OFF.
This BroadcastReceiver listens to the simstatechange
always when BT is ON. Signal strength and service
availability values sent to connected devices depending
on the sim state.

Change-Id: I4f903afd07e2719cee2ee093edfdc5526fd865e4

Delay connection time-out.

This change delays the auto connect of audio profiles
to make sure that all the clients of the Bluetooth audio
services are ready. This handles a corner case of BT
OFF-ON and auto connect.

Change-Id: I14b5ebc37617847f87387dfee501b59427c9d14a

Query phone states when SLC is established

Handling QUERY_PHONE_STATE_AT_SLC in
connectedBase class so that it would be
made visible to Connected, AudioConnecting,
AudioOn, AudioDisconnecting states.

CRs-Fixed: 2207073
Change-Id: Ib9cbd0297703d77e5af161705177e0eff5234c7e