HFP: Sending service availablity and signal strength in +CIND response
When the BT is turned ON, the +CIND response would be updated
with no network available and no signal, even though if the call
is active/held/in-progress. This would cause some of the carkits/
handsfree devices to go into bad state.

During BT turn ON, when the remote device is getting connected,
and also if the call is active/held/in-progress. The +CIND response
is updated with network available and signal strength.

Also when the RD is in connecting state, the change in the call
states are updated by processCallState API.

Change-Id: I755ebb71692f5553d23f75127cafdced7967d57a
CRs-Fixed: 2243724