HFP:Inband ring is enabled only for one device connected
When the inbandringing is supported and the
"persist.bluetooth.disableinbandringing" is not set, the
inband ringtone is heard only when one device is
connected.

If more than one device is connected, during runtime
mInbandRingingRuntimeDisable is set to true and
+BSIR = 0 is sent to the connected devices.

By default, inband ringtone is disabled and can be
enabled by setting the property
"persist.bluetooth.disableinbandringing" to false.

CRs-Fixed: 2294689
Change-Id: If532d61d0cec08dd1559770f5a6abd08e37045cf