MAP: [IOT] Update MAS-0 folder name case to small letters
Precondition:
MAP Connection for MAS-0 with MNS connected.

Usecase:
1) Establish MAP MAS session for SMS.
2) Send message from carkit and wait for sending success
   event report.
3) Observe if sms in CarKit is updated as sent or not.
4) Similar case with read status.

Issue:
SMS in CarKit is not updated as sent.

Fix:
This change updates case of mandatory folder name of MAS-0
account.

Root Cause:
Remote carkit F_Ecosport is case sensitive towards folder
name. So folder name sent in Capital letters is not processed
by this carkit as case of foldername in Folderlisting and Event
Report is not same (telecom/msg/sent for FolderListing and
telecom/msg/SENT for EventReport).

Change-Id: I250cefc0516a16d573ee5f6d33aad076bfb98c16
CRs-Fixed: 2271014