HFP: Check the status of the SIM in DSDS mode
In DSDS mode, if only one sim is present in either
of the sim slots, we used to treat it as sim not
loaded. As a result of which service and signal
strength indicators were set to zero.

The change includes to get both sim states at the
BT turn ON. Later, the change in sim state is
updated as per the intents received. Even if one
sim is loaded, treat as sim is loaded.

The bluetooth process tries to catch the
exception while accessing the telephony APIs for
getting sim states and also while registering
for signal strength notifications.

CRs-Fixed: 2310992
Change-Id: Ia3293837b07d49b8e971b90e4d07099371cb91d2