HFP: Send call indicators to remote in serialized way
When A2DP is playing and a call is made back to back,
the call indicators are not sent correctly to the remote
device and the remote device would go into bad state.

To avoid this race condition, the call indicators are
cached before A2DP is suspended and later the indicators
are  sent in the serialized way to the remote device
when the call is terminated immediately.

CRs-Fixed: 2289855
Change-Id: I6feb288133f4f085d1b11b60d62b24df5c65572a