PBAP: Handle Version Upgrade Notification on completing pairing
In this change, PBAP PCE SDP Search is moved to stack and will
be done during pairing itself. On Bond state BONDED, logic to
check if notification is required for version upgrade is
handled. This is triggered 500ms post state is changed to
BOND_BONDED so that config file is updated in stack by that time.

Change-Id: Ib3cecdcd4e3ef1851ea417a400ca94aae3bdf789
CRs-Fixed: 2380390