A2DP: Send disconnectA2dp explicitly for incoming connection
Send disconnect explicitly for incoming connection for new device
if the max connections are reached. For previously connected
device if not unbonded, a2dp sm will be present which takes care
of sending disconnectA2dp if the connection is not allowed. If
BT is toggled a2dp sm will be created only if the connection is
allowed. So disconnect has to be called if sm is not present for
the device.

CRs-fixed: 2446784
Change-Id: Ia424b1cbe92fd9ad44cb476ae4622cfaa7436ea6