TWSP-HFP: Stop VR while SCO connection only setup with inactive earbud
If active EB is out of ear, inactive EB is in ear. After VR is started,
SCO connection is established only with the in ear EB. While stop VR,
requesting device will set to be active device, so that disconnect audio
request no chance to send to the inactive in ear EB and SCO connection
can not be terminated.

Need set the inactive and audio connected EB as requesting device instead
of active but out of earbud EB while stop VR.

Change-Id: Ia2173ee2ec3218a59cc424f8968187eaf54ca3ac
CRs-fixed: 2512072