A2dpSink: Accept Absolute volume from non-streaming source.
When non-streaming source device sent absolute volume with some
value(x%), sink device is not accepting with same value(x%), it
is accepting with value(y%) which it has currently streaming(y%)
and also sent changed rsp with same(y%). Due to this, non-streaming
source device is continuousely sending absolute volume with +1/-1
for matching and till '0', if streaming would be on 100% with active
streaming device.

This change would take care of accepting the non-streaming source
device sent absolute volume(x%) at sink end, and respond with Sink
volume(y%).

CRs-Fixed: 2675377
Change-Id: I896af989453d4ddeb61361d7fcf42089d22074a1