A2dpSink: Accept Absolute volume from non-streaming source.
When non-streaming source device sent absolute volume with some
value(x%), sink device is not accepting with same value(x%), it is
accepting with value(y%) which it has currently streaming(y%) and
also sent changed rsp with same(y%). Due to this, non-streaming
source device is continuousely sending absolute volume with +1/-1
for matching and till '0', if streaming would be on 100% with active
streaming device.

This change would take care of accepting the non-streaming source
device sent absolute volume(x%) at sink end, and respond with Sink
volume(y%).

CRs-Fixed: 2675421
Change-Id: Ib94ad60bbf4de27f2216c4ffb42df26a0c0ce476