Send BSIR:0 as soon as one more device is connected
After sending BSIR:0 when second device gets connected,
in-band ringtone enabled flag is set to false. This
leads to not sending BSIR:0 when more devices are
connected. Fix the condition to check if in-band
ringtone is enabled in overlay and property to be
able to send BSIR:0 after SLC connection.

CRs-Fixed: 2689681
Change-Id: I382d9b3bca77129347249e2d3a83c84def1044d1