HFP: Suspended A2dp flag when call is in progress is resetted when Bt is turned off
When call is in received, we suspend the A2dp to audio manager for the voice
connection. when call is in over, we make the A2dp suspend status false.
When BT gets turned off during the call, we are not making it false and so
when the BT is connected again, the value A2dp remained suspended. To avoid
this, when BT gets turned off, the A2dp suspend value is made false.

CRs-Fixed: 2718036

Change-Id: Id22389b114c04bef0a3365792b616b3e06961899