Decouple screen and location checks in ScanManager.
This separates checks in the BT LE ScanManager for the screen off
and location off conditions. Previously the two were connected so
that location is checked when the screen is turned off, and vice
versa.

This also adds a hasDisavowedLocation property to ScanClient, and
includes it in checks but the property is not yet populated and
is always false currently.

Bug: 183203469
Test: atest GattServiceTest

Not a Contribution

Change-Id: I85cf6db4df160f2d6b0247db618e0d70baf5de6e
(cherry picked from commit 79359d07dd96cac86e1357c0bac92b67379c226a)