Bluetooth: AOSP changes related to permissions.
Add AttributionSource to BT methods.

This adds attribution source as a parameter to bluetooth methods,
matching aidl changes also made in this topic.

This is now required to allow the app ops for the new
bluetooth permissions (BLUETOOTH_CONNECT, BLUETOOTH_ADVERTISE,
and BLUETOOTH_SCAN) to be noted.

Update Bluetooth API annotations.

* BLUETOOTH_ADVERTISE: Required to be able to advertise to nearby
                       Bluetooth devices.
* BLUETOOTH_CONNECT:   Allows applications to connect to paired
                       bluetooth devices.
* BLUETOOTH_SCAN:      Required to be able to discover and pair
                       nearby Bluetooth devices.

Change-Id: Idbcd7f82e86680540709924c76955bc58b9201c8
( Not a contribution )